+++ docs/changelog.md
@@ -25,6 +25,9 @@ permalink: /changelog/
 * **feature-search**
   * Added `newPrivateTabSearch` `NewTabSearchUseCase`
 
+* **feature-session**
+  * Added ability to `removePrivateSessions` from a `SessionManager`
+  
 * **feature-toolbar**
   * Added ability to color parts of the domain (e.g. [registrable domain](https://url.spec.whatwg.org/#host-registrable-domain)) by providing a `UrlRenderConfiguration`: